+++ data/highscore7.txt
@@ -1,5 +1,5 @@
 1	Iwin	43214
-2	Joseph	3432
-3	Abal	500
-4	Fell	483
-5	Bite	400
+2	ZMS	5339
+3	ZMS	400
+4	Joseph	3432
+5	Abal	500
