+++ src/highscore_screen.cpp
@@ -1,6 +1,113 @@
 #include "defs.h"
 #include "highscore_screen.h"
 
+
+void get_scores(int diffuculty){
+		
+		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 7"};
+		
+		ifstream ist{"data/highscore"+to_string(diffuculty)+".txt"};
+		if(!ist)error("cant open");
+
+		vector<List>highscore;
+		int place;
+		string name;
+		int score;
+
+		while(ist>>place>>name>>score){
+			highscore.push_back(List{place,name,score});
+}
+
+
+ist.close();
+		
+		ostringstream one;
+		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
+		Text first{Point{50,70},one.str()};
+		first.set_font_size(30);
+		first.set_color(Color::yellow);
+		scores.attach(first);
+		
+		ostringstream two;
+		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
+		Text sec{Point{50,145},two.str()};
+		sec.set_font_size(30);
+		sec.set_color(Color{18});
+		scores.attach(sec);
+		
+		ostringstream three;
+		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
+		Text third{Point{50,220},three.str()};
+		third.set_font_size(30);
+		third.set_color(Color{74});
+		scores.attach(third);
+		
+		ostringstream four;
+		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
+		Text fourth{Point{50,295},four.str()};
+		fourth.set_font_size(30);
+		scores.attach(fourth);
+		
+		ostringstream five;
+		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
+		Text fifth{Point{50,370},five.str()};
+		fifth.set_font_size(30);
+		scores.attach(fifth);
+		
+		scores.wait_for_button();
+	}
+	
+void score_sorter(int f, string nam, int diffuculty){
+	
+	ifstream ist{"data/highscore"+to_string(diffuculty)+".txt"};
+	if(!ist)error("cant open");
+
+	vector<List>highscore;
+		int place;
+		string name;
+		int score;
+
+	while(ist>>place>>name>>score){
+		highscore.push_back(List{place,name,score});
+	}
+	
+	if(f>highscore[4].score){
+	highscore[4].score=f;
+	highscore[4].name=nam;
+
+	if(f>highscore[3].score){
+		highscore[4].score=highscore[3].score;
+		highscore[3].score=f;
+		highscore[4].name=highscore[3].name;
+		highscore[3].name=nam;
+}
+		if(f>highscore[2].score){
+			highscore[3].score=highscore[2].score;
+			highscore[2].score=f;
+			highscore[3].name=highscore[2].name;
+			highscore[2].name=nam;
+}
+			if(f>highscore[1].score){
+				highscore[2].score=highscore[1].score;
+				highscore[1].score=f;
+				highscore[2].name=highscore[1].name;
+				highscore[1].name=nam;
+}
+				if(f>highscore[0].score){
+					highscore[1].score=highscore[0].score;
+					highscore[0].score=f;
+					highscore[1].name=highscore[0].name;
+					highscore[0].name=nam;
+}
+}
+	
+	ofstream ost{"data/highscore"+to_string(diffuculty)+".txt"};
+	ost.trunc;
+
+	for(int i=0; i<highscore.size(); ++i)
+		ost<<highscore[i].place<<'\t'<<highscore[i].name<<'\t'<<highscore[i].score<<'\n';
+}
+
 highscore_screen::highscore_screen(Point xy,int w,int h,const string& title, int& difficulty)
 :Window{xy,w,h,title},
 
@@ -50,260 +157,30 @@ diff{difficulty}
 		attach(game7);
 	}
 
+
 	void highscore_screen::high3()
 	{
-		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 3"};
-		
-		ifstream ist{"data/highscore3.txt"};
-		if(!ist)error("cant open");
-
-		vector<List>highscore;
-		int place;
-		string name;
-		int score;
-
-		while(ist>>place>>name>>score){
-			highscore.push_back(List{place,name,score});
-}
-
-ist.close();
-		
-		ostringstream one;
-		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
-		Text first{Point{50,70},one.str()};
-		first.set_font_size(30);
-		scores.attach(first);
-		
-		ostringstream two;
-		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
-		Text sec{Point{50,145},two.str()};
-		sec.set_font_size(30);
-		scores.attach(sec);
-		
-		ostringstream three;
-		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
-		Text third{Point{50,220},three.str()};
-		third.set_font_size(30);
-		scores.attach(third);
-		
-		ostringstream four;
-		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
-		Text fourth{Point{50,295},four.str()};
-		fourth.set_font_size(30);
-		scores.attach(fourth);
-		
-		ostringstream five;
-		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
-		Text fifth{Point{50,370},five.str()};
-		fifth.set_font_size(30);
-		scores.attach(fifth);
-		
-		
-		
-		scores.wait_for_button();
+		get_scores(3);
 	}
 	
 	void highscore_screen::high4()
 	{
-		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 4"};
-		ifstream ist{"data/highscore4.txt"};
-		if(!ist)error("cant open");
-
-		vector<List>highscore;
-		int place;
-		string name;
-		int score;
-
-		while(ist>>place>>name>>score){
-			highscore.push_back(List{place,name,score});
-}
-
-ist.close();
-		
-		ostringstream one;
-		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
-		Text first{Point{50,70},one.str()};
-		first.set_font_size(30);
-		scores.attach(first);
-		
-		ostringstream two;
-		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
-		Text sec{Point{50,145},two.str()};
-		sec.set_font_size(30);
-		scores.attach(sec);
-		
-		ostringstream three;
-		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
-		Text third{Point{50,220},three.str()};
-		third.set_font_size(30);
-		scores.attach(third);
-		
-		ostringstream four;
-		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
-		Text fourth{Point{50,295},four.str()};
-		fourth.set_font_size(30);
-		scores.attach(fourth);
-		
-		ostringstream five;
-		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
-		Text fifth{Point{50,370},five.str()};
-		fifth.set_font_size(30);
-		scores.attach(fifth);
-		
-		scores.wait_for_button();
+		get_scores(4);
 	}
 	
 	void highscore_screen::high5()
 	{
-		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 5"};
-		
-		ifstream ist{"data/highscore5.txt"};
-		if(!ist)error("cant open");
-
-		vector<List>highscore;
-		int place;
-		string name;
-		int score;
-
-		while(ist>>place>>name>>score){
-			highscore.push_back(List{place,name,score});
-}
-
-ist.close();
-		
-		ostringstream one;
-		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
-		Text first{Point{50,70},one.str()};
-		first.set_font_size(30);
-		scores.attach(first);
-		
-		ostringstream two;
-		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
-		Text sec{Point{50,145},two.str()};
-		sec.set_font_size(30);
-		scores.attach(sec);
-		
-		ostringstream three;
-		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
-		Text third{Point{50,220},three.str()};
-		third.set_font_size(30);
-		scores.attach(third);
-		
-		ostringstream four;
-		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
-		Text fourth{Point{50,295},four.str()};
-		fourth.set_font_size(30);
-		scores.attach(fourth);
-		
-		ostringstream five;
-		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
-		Text fifth{Point{50,370},five.str()};
-		fifth.set_font_size(30);
-		scores.attach(fifth);
-		
-		scores.wait_for_button();
+		get_scores(5);
 	}
 	
 	void highscore_screen::high6()
 	{
-		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 6"};
-		
-		ifstream ist{"data/highscore6.txt"};
-		if(!ist)error("cant open");
-
-		vector<List>highscore;
-		int place;
-		string name;
-		int score;
-
-		while(ist>>place>>name>>score){
-			highscore.push_back(List{place,name,score});
-}
-
-ist.close();
-		
-		ostringstream one;
-		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
-		Text first{Point{50,70},one.str()};
-		first.set_font_size(30);
-		scores.attach(first);
-		
-		ostringstream two;
-		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
-		Text sec{Point{50,145},two.str()};
-		sec.set_font_size(30);
-		scores.attach(sec);
-		
-		ostringstream three;
-		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
-		Text third{Point{50,220},three.str()};
-		third.set_font_size(30);
-		scores.attach(third);
-		
-		ostringstream four;
-		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
-		Text fourth{Point{50,295},four.str()};
-		fourth.set_font_size(30);
-		scores.attach(fourth);
-		
-		ostringstream five;
-		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
-		Text fifth{Point{50,370},five.str()};
-		fifth.set_font_size(30);
-		scores.attach(fifth);
-		
-		scores.wait_for_button();
+		get_scores(6);
 	}
 	
 	void highscore_screen::high7()
 	{
-		Simple_window scores{Point{200,200},400,400,"Highscore, Difficulty: 7"};
-		
-		ifstream ist{"data/highscore7.txt"};
-		if(!ist)error("cant open");
-
-		vector<List>highscore;
-		int place;
-		string name;
-		int score;
-
-		while(ist>>place>>name>>score){
-			highscore.push_back(List{place,name,score});
-}
-
-ist.close();
-		
-		ostringstream one;
-		one<<highscore[0].place<<". "<<highscore[0].name<<'-'<<highscore[0].score<<endl;
-		Text first{Point{50,70},one.str()};
-		first.set_font_size(30);
-		scores.attach(first);
-		
-		ostringstream two;
-		two<<highscore[1].place<<". "<<highscore[1].name<<'-'<<highscore[1].score<<endl;
-		Text sec{Point{50,145},two.str()};
-		sec.set_font_size(30);
-		scores.attach(sec);
-		
-		ostringstream three;
-		three<<highscore[2].place<<". "<<highscore[2].name<<'-'<<highscore[2].score<<endl;
-		Text third{Point{50,220},three.str()};
-		third.set_font_size(30);
-		scores.attach(third);
-		
-		ostringstream four;
-		four<<highscore[3].place<<". "<<highscore[3].name<<'-'<<highscore[3].score<<endl;
-		Text fourth{Point{50,295},four.str()};
-		fourth.set_font_size(30);
-		scores.attach(fourth);
-		
-		ostringstream five;
-		five<<highscore[4].place<<". "<<highscore[4].name<<'-'<<highscore[4].score<<endl;
-		Text fifth{Point{50,370},five.str()};
-		fifth.set_font_size(30);
-		scores.attach(fifth);
-		
-		scores.wait_for_button();
+		get_scores(7);
 	}
 	
 	void highscore_screen::gamefun3(){
